Implement LetterListFragment
As with activities, you need to inflate the layout and bind individual views. There are just a few minor differences when working with the fragment lifecycle. We'll walk you through the process for setting up the LetterListFragment, and then you'll get the chance to do the same for WordListFragment.

To implement view binding in LetterListFragment, you first need to get a nullable reference to FragmentLetterListBinding. Binding classes like this are generated by Android Studio for each layout file, when the viewBinding property is enabled under the buildFeatures section of the build.gradle file. You just need to assign properties in your fragment class for each view in the FragmentLetterListBinding.

The type should be FragmentLetterListBinding? and it should have an initial value of null. Why make it nullable? Because you can't inflate the layout until onCreateView() is called. There's a period of time in-between when the instance of LetterListFragment is created (when its lifecycle begins with onCreate()) and when this property is actually usable. Also keep in mind that fragments' views can be created and destroyed several times throughout the fragment's lifecycle. For this reason you also need to reset the value in another lifecycle method, onDestroyView().